% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr.SN.fit.R
\name{clr.SN.fit}
\alias{clr.SN.fit}
\title{Fitting the skew-normal distribution to CLR-transformed RNA-Seq data}
\usage{
clr.SN.fit(data)
}
\arguments{
\item{data}{A table of CLR-transformed count data, where genes/transcripts on the rows and
samples on columns.}
}
\value{
\item{mu}{The maximum likelihood estimate of the mean parameter.}
\item{se.mu}{The standard error of the maximum likelihood estimate of mu.}
\item{z.mu}{The Wald statistic for mu.}
\item{p.mu}{The p-value of the Wald statistic for mu.}
\item{sigma}{The maximum likelihood estimate of the standard deviation parameter.}
\item{se.sigma}{The standard error of the maximum likelihood estimate of sigma.}
\item{z.sigma}{The Wald statistic for sigma.}
\item{p.sigma}{The p-value of the Wald statistic for sigma.}
\item{gamma}{The maximum likelihood estimate of the skewness parameter.}
\item{se.gamma}{The standard error of the maximum likelihood estimate of gamma.}
\item{z.gamma}{The Wald statistic for gamma.}
\item{p.gamma}{The p-value of the Wald statistic for gamma.}
}
\description{
Estimate the mean, standard deviation, and skewness parameters of
the skew-normal distribution using centered log-ratio (CLR)
transformed RNA-Seq data.
}
\examples{
 library(clrDV)
 data(clrCounts1)
 clr.SN.fit(clrCounts1[1:2, ])
 clr.SN.fit(clrCounts1[1, ])



}
