% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clrCounts1.R
\docType{data}
\name{clrCounts1}
\alias{clrCounts1}
\title{Simulated CLR-transformed counts table}
\format{
A data frame contains CLR-transformed counts, with 500 genes and 500 samples.
\describe{
\item{clrCounts1}{CLR-transformed counts table, 500*500.}
}
}
\usage{
clrCounts1
}
\description{
A simulated clr-transformed counts table with genes on rows and samples on columns for one condition.
}
\keyword{datasets}
