% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clrCounts2.R
\docType{data}
\name{clrCounts2}
\alias{clrCounts2}
\title{Simulated DV CLR-transformed counts table}
\format{
A data frames contains CLR-transformed counts, with 1000 genes and 200 samples per group.
Ten percent of these genes are DV genes (gene1 to gene100).
\describe{
\item{clrCounts2}{DV CLR-transformed counts table, 1000*400.}
}
}
\usage{
clrCounts2
}
\description{
A simulated DV CLR-transformed counts table with genes on rows and samples on columns for two groups.
}
\keyword{datasets}
