% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin.plot.clrDV.R
\name{violin.plot.clrDV}
\alias{violin.plot.clrDV}
\title{Violin plots}
\usage{
violin.plot.clrDV(
  data = NULL,
  name.gene = NULL,
  group = NULL,
  group.names = NULL,
  xlab = "CLR-transformed count",
  ylab = "Condition"
)
}
\arguments{
\item{data}{A CLR-transformed count table.}

\item{name.gene}{Gene/transcript name.}

\item{group}{A vector specifying the group labels of the data.}

\item{group.names}{A vector specifying the group names.}

\item{xlab}{Name of the x-axis.}

\item{ylab}{Name of the y-axis.}
}
\description{
Produce violin plots of CLR-transformed count data.
}
\examples{
  data(clrCounts2)
  group0 <- c(rep(0, 200), rep(1, 200))
  group1 <- c(rep(0, 200), rep(1, 100), rep(2, 100))
  violin.plot.clrDV(data = clrCounts2, "gene2", group = group0,
                group.names = c("control", "case"))
  violin.plot.clrDV(data = clrCounts2, "gene2", group = group1,
                group.names = c("control", "case1", "case2"))



}
