% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top.DV.genes.R
\name{top.DV.genes}
\alias{top.DV.genes}
\title{Top-ranked genes that show differential variability}
\usage{
top.DV.genes(clrDV_result, top = "abs.", n = 10, full.table = F)
}
\arguments{
\item{clrDV_result}{An output from clrDV() function.}

\item{top}{large/small/abs. Large for SD ratio > 1; small for SD ratio < 1;
abs. for ranking the DV genes by the absolute value of log2 SD ratio.}

\item{n}{The threshold rank.}

\item{full.table}{Default is FALSE. If full.table = TRUE, the maximum likelihood
estimates of parameters of skew-normal distribution for group 1 and group 2 will be included.}
}
\value{
\item{DV}{The difference of the estimated standard deviation between group 2 and group 1 (sigma2 - sigma1).}
\item{se}{The standard error of DV.}
\item{z-value}{The observed Wald statistic value.}
\item{p-value}{The unadjusted p-value of the Wald test.}
\item{adjusted p-value}{The p-value of the Wald test adjusted using the Benjamini-Yekutieli procedure.}
\item{sd_ratio}{The ratio of sigma2 and sigma1 (sigma2/sigma1).}
\item{LFC}{log2 of the sd_ratio.}
}
\description{
Extract top-ranked differential variability (DV) genes. These genes are ranked
using the SD ratio (treatment vs. control) of the CLR-transformed count.
}
\examples{
   library(clrDV)
   data("clrCounts2")
   group0 <- c(rep(0, 200), rep(1, 200))
   simu_clrDV_test <- clrDV(clrCounts2, group0)
   top.DV.genes(simu_clrDV_test, top = "large", n = 5)
   top.DV.genes(simu_clrDV_test, top = "small", n = 5)
   top.DV.genes(simu_clrDV_test, top = "abs.", n = 5)
   top.DV.genes(simu_clrDV_test, top = "abs.", n = 5, full.table = TRUE)
   dim(top.DV.genes(simu_clrDV_test, top = "abs.", n = Inf))


}
